pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "docusign-agreements-downloader"
version = "0.1.0"
description = "Download DocuSign agreements (eSignature envelopes) and documents to the filesystem."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name="Your Team"}]
dependencies = [
  "httpx>=0.27.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.3.0",
  "PyJWT[crypto]>=2.8.0",
  "tenacity>=9.0.0",
  "typer>=0.12.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0",
  "respx>=0.21.0",
  "freezegun>=1.5.0",
  "ruff>=0.6.0",
  "mypy>=1.10.0",
  "types-PyJWT>=1.7.1.20240417",
]

[project.scripts]
dsa = "docusign_agreements_downloader.cli:app"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.coverage.run]
branch = true
source = ["docusign_agreements_downloader"]

[tool.coverage.report]
show_missing = true
fail_under = 90
